pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["kingahmadr <ahmadnromiz@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
flask-login = "^0.6.3"
python-dotenv = "^1.0.1"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.0.7"
flasgger = "^0.9.7.1"
werkzeug = "^3.0.4"
jwt = "^1.3.1"
pyjwt = "^2.9.0"
pytest = "^8.3.3"
pytest-cov = "^5.0.0"
pytest-mock = "^3.14.0"
pymysql = "^1.1.1"
flask-jwt-extended = "^4.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
